[REFACTOR] 서비스 분리
